refactor: Move CustomTooltip to separte file
